+++ packages/game-core/src/index.ts
@@ -1,4 +1,5 @@
 export * from './types/puzzleTypes';
 export * from './utils/geometry/puzzleGeometry';
+export * from './utils/geometry/IncrementalCutter';
 export * from './utils/score/ScoreCalculator';
 export * from './utils/score/RotationEfficiencyCalculator';